requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.1.post1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2022.5.18.1
cfscrape==2.1.1
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
distlib==0.3.4
Django==4.0.1
django-cors==0.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
exponent-server-sdk==2.0.0
filelock==3.6.0
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
mailjet-rest==1.3.4
Markdown==3.3.6
markdownify==0.11.6
multidict==6.0.2
mysqlclient==2.1.1
nltk==3.6.5
numpy==1.21.4
oauth2client==3.0.0
pandas==1.3.5
Pillow==9.1.1
platformdirs==2.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.27.1
requests-toolbelt==0.7.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
tzdata==2021.5
tzlocal==4.2
uplink==0.9.7
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
whitenoise==5.2.0
yarl==1.7.2
zipp==3.8.0
